Use the gradle toolchain for defining the java version. Add a gestalt-test project to hold the tests for records, as the gestalt-core tests are now jdk 11. Support jacoco code coverage reports. Fix a bug with gitConfigSourceBuilder not persisting the tags. For the empty set of tags, use a single static instance.
