+++ build-logic/src/main/kotlin/gestalt.java-common-conventions.gradle.kts
@@ -14,12 +14,9 @@ repositories {
 
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-}
-
-tasks.compileJava {
-    options.compilerArgs.addAll(arrayOf("--release", "11"))
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(11))
+    }
 }
 
 dependencies {
