+++ build-logic/src/main/kotlin/gestalt.kotlin-test-conventions.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
-    jacoco
     id("gestalt.kotlin-common-conventions")
+    jacoco
 }
 
 dependencies {
