--- gestalt-core/src/test/java/org/github/gestalt/config/test/classes/Person2.java
@@ -1,4 +1,0 @@
-package org.github.gestalt.config.test.classes;
-
-public record Person2(Integer id, String name) {
-}
