--- gestalt-core/src/test/java/org/github/gestalt/config/test/classes/Person3.java
@@ -1,6 +1,0 @@
-package org.github.gestalt.config.test.classes;
-
-import org.github.gestalt.config.annotations.Config;
-
-public record Person3(@Config(path = "test") Integer id, String name) {
-}
