+++ gestalt-sample-java-latest/build.gradle.kts
@@ -10,9 +10,10 @@ repositories {
     mavenCentral()
 }
 
-tasks.compileTestJava {
-    sourceCompatibility = JavaVersion.VERSION_17.toString()
-    targetCompatibility = JavaVersion.VERSION_17.toString()
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(17))
+    }
 }
 
 dependencies {
