+++ gestalt-test/build.gradle.kts
@@ -2,9 +2,14 @@ plugins {
     id("gestalt.java-library-conventions")
     id("gestalt.java-test-conventions")
     id("gestalt.java-code-quality-conventions")
-    id("gestalt.java-publish-conventions")
 }
 
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(17))
+    }
+}
 dependencies {
+    testImplementation(project(":gestalt-core"))
     testImplementation(libs.slf4j.simple)
 }
