+++ gestalt-test/src/test/java/org/github/gestalt/config/test/classes/DBInfo.java
@@ -1,0 +1,31 @@
+package org.github.gestalt.config.test.classes;
+
+public class DBInfo {
+    private int port;
+    private String uri;
+    private String password;
+
+    public int getPort() {
+        return port;
+    }
+
+    public void setPort(int port) {
+        this.port = port;
+    }
+
+    public String getUri() {
+        return uri;
+    }
+
+    public void setUri(String uri) {
+        this.uri = uri;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
