+++ gestalt-test/src/test/java/org/github/gestalt/config/test/classes/DBPool.java
@@ -1,0 +1,13 @@
+package org.github.gestalt.config.test.classes;
+
+public class DBPool {
+    public int maxTotal;
+    public int maxPerRoute;
+    public int validateAfterInactivity;
+    public int keepAliveTimeoutMs = 6000;
+    public int idleTimeoutSec = 10;
+    public float defaultWait = 33.0F;
+
+    public DBPool() {
+    }
+}
