+++ gestalt-test/src/test/java/org/github/gestalt/config/test/classes/Person.java
@@ -1,0 +1,5 @@
+package org.github.gestalt.config.test.classes;
+
+public record Person(String name, Integer id) {
+}
+
