+++ gestalt-test/src/test/java/org/github/gestalt/config/test/classes/PersonAnnotations.java
@@ -1,0 +1,6 @@
+package org.github.gestalt.config.test.classes;
+
+import org.github.gestalt.config.annotations.Config;
+
+public record PersonAnnotations(String name, @Config(path = "identity", defaultVal = "1234") Integer id) {
+}
