+++ settings.gradle.kts
@@ -1,10 +1,13 @@
 rootProject.name = "gestalt"
 include(
-    "gestalt-benchmark", "gestalt-core", "gestalt-hocon", "gestalt-json", "gestalt-git", "gestalt-kotlin", "gestalt-kodein-di",
-    "gestalt-koin-di", "gestalt-s3", "gestalt-sample", "gestalt-sample-java-latest", "gestalt-toml", "gestalt-yaml"
+    "gestalt-core", "gestalt-hocon", "gestalt-json", "gestalt-git", "gestalt-kotlin", "gestalt-kodein-di",
+    "gestalt-koin-di", "gestalt-s3", "gestalt-toml", "gestalt-yaml"
 )
 
 includeBuild("build-logic")
+
+// testing utility projects
+include("code-coverage-report", "gestalt-benchmark", "gestalt-sample", "gestalt-sample-java-latest", "gestalt-test")
 
 pluginManagement {
     repositories {
